corrigiendo bug al guardar precio de producto

+++ database/migrations/2021_12_12_204635_create_productos_table.php
@@ -17,8 +17,7 @@ class CreateProductosTable extends Migration
             $table->id();
             $table->string('nombre');
             $table->string('descripcion');
-            $table->float('precio');
-
+            $table->decimal('precio');
             $table->timestamps();
         });
     }
